Move self-test message schema out of dropbot_controller consts
consts.py held a mid-file json import, a TestEvent namespace and the
create_test_progress_message helper between two topic blocks. They are
the SELF_TESTS_PROGRESS payload contract, so they now live in
dropbot_controller/models/self_tests.py; the mixin service, the
frontend task and the mock controller import from there.
