+++ mock_dropbot_controller/consts.py
@@ -11,8 +11,10 @@ from dropbot_controller.consts import (
     SET_REALTIME_MODE, RUN_ALL_TESTS, TEST_VOLTAGE,
     TEST_ON_BOARD_FEEDBACK_CALIBRATION, TEST_SHORTS, TEST_CHANNELS,
     CHIP_CHECK, SELF_TEST_CANCEL, DETECT_DROPLETS, CHANGE_SETTINGS,
-    HARDWARE_DEFAULT_VOLTAGE, HARDWARE_DEFAULT_FREQUENCY, TestEvent,
-    create_test_progress_message
+    HARDWARE_DEFAULT_VOLTAGE, HARDWARE_DEFAULT_FREQUENCY,
+)
+from dropbot_controller.models.self_tests import (
+    TestEvent, create_test_progress_message,
 )
 
 from electrode_controller.consts import ELECTRODES_STATE_APPLIED
